+++ src/EVM/Solvers.hs
@@ -265,13 +265,13 @@ getModel :: SolverInstance -> CexVars -> IO SMTCex
 getModel inst cexvars = do
   -- get an initial version of the model from the solver
   initialModel <- getRaw
-  -- get concrete values for each buffers max read index
-  hints <- capHints <$> queryMaxReads (getValue inst) cexvars.buffers
   -- check the sizes of buffer models and shrink if needed
   if bufsUsable initialModel
-  then do
-    pure (mkConcrete initialModel)
-  else mkConcrete . snd <$> runStateT (shrinkModel hints) initialModel
+  then pure (mkConcrete initialModel)
+  else do
+    -- get concrete values for each buffers max read index
+    hints <- capHints <$> queryMaxReads (getValue inst) cexvars.buffers
+    mkConcrete . snd <$> runStateT (shrinkModel hints) initialModel
   where
     getRaw :: IO SMTCex
     getRaw = do
@@ -321,7 +321,10 @@ getModel inst cexvars = do
           put model
         "unsat" -> do
           liftIO $ checkCommand inst "(pop 1)"
-          shrinkBuf buf (if hint == 0 then hint + 1 else hint * 2)
+          let nextHint = if hint == 0 then 1 else hint * 2
+          if nextHint < hint
+            then pure () -- overflow
+            else shrinkBuf buf nextHint
         e -> internalError $ "Unexpected solver output: " <> (T.unpack e)
 
     -- Collapses the abstract description of a models buffers down to a bytestring