+++ src/EVM/Types.hs
@@ -588,12 +588,14 @@ data PartialExec
   | MaxIterationsReached  { pc :: Int, addr :: Expr EAddr }
   | JumpIntoSymbolicCode  { pc :: Int, jumpDst :: Int }
   | CheatCodeMissing      { pc :: Int, selector :: FunctionSelector }
+  | BranchTooDeep         { pc :: Int}
   deriving (Show, Eq, Ord)
 
 -- | Effect types used by the vm implementation for side effects & control flow
 data Effect t s where
   Query :: Query t s -> Effect t s
   RunBoth :: RunBoth s -> Effect Symbolic s
+  RunAll :: RunAll s -> Effect Symbolic s
 deriving instance Show (Effect t s)
 
 -- | Queries halt execution until resolved through RPC calls or SMT queries
@@ -608,6 +610,10 @@ data Query t s where
 -- | Execution could proceed down one of two branches
 data RunBoth s where
   PleaseRunBoth    :: Expr EWord -> (Bool -> EVM Symbolic s ()) -> RunBoth s
+
+-- | Execution could proceed down one of several branches
+data RunAll s where
+  PleaseRunAll    :: Expr EWord -> [W256] -> (W256 -> EVM Symbolic s ()) -> RunAll s
 
 -- | The possible return values of a SMT query
 data BranchCondition = Case Bool | Unknown
@@ -640,6 +646,11 @@ instance Show (RunBoth s) where
   showsPrec _ = \case
     PleaseRunBoth _ _ ->
       (("<EVM.RunBoth: system running both paths") ++)
+
+instance Show (RunAll s) where
+  showsPrec _ = \case
+    PleaseRunAll _ _ _ ->
+      (("<EVM.RunAll: system running all paths") ++)
 
 -- | The possible result states of a VM
 data VMResult (t :: VMType) s where
@@ -682,6 +693,7 @@ data VM (t :: VMType) s = VM
   , freshVar       :: Int
   -- ^ used to generate fresh symbolic variable names for overapproximations
   --   during symbolic execution. See e.g. OpStaticcall
+  , exploreDepth    :: Int
   }
   deriving (Generic)
 
@@ -870,8 +882,8 @@ class VMOps (t :: VMType) where
   whenSymbolicElse :: EVM t s a -> EVM t s a -> EVM t s a
 
   partial :: PartialExec -> EVM t s ()
-  branch :: Expr EWord -> (Bool -> EVM t s ()) -> EVM t s ()
-  manySolutions :: Expr EWord -> Int -> (Maybe W256 -> EVM t s ()) -> EVM t s ()
+  branch :: Maybe Int -> Expr EWord -> (Bool -> EVM t s ()) -> EVM t s ()
+  manySolutions :: Maybe Int -> Expr EWord -> Int -> (Maybe W256 -> EVM t s ()) -> EVM t s ()
 
 -- Bytecode Representations ------------------------------------------------------------------------