+++ CHANGELOG.md
@@ -46,6 +46,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   print some warnings related to zero-address dereference and to print `hemv test`'s
   output in case of failure
 - Simple test cases for the CLI
+- Allow limiting the branch depth and width limitation via --max-depth and --max-width
 
 ## Fixed
 - We now try to simplify expressions fully before trying to cast them to a concrete value
@@ -67,6 +68,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Equivalence checking was incorrectly assuming that overapproximated values
   were sequentially equivalent. We now distinguish these symbolic values with
   `A-` and `B-`
+- Buffer of all zeroes was interpreted as an empty buffer during parsing SMT model.
+  The length of the buffer is now properly taken into account.
+- It was possible to enter an infinite recursion when trying to shrink a buffer found by
+  the SMT solver. We now properly detect that it is not possible to shrink the buffer.
 
 ## Changed
 - Warnings now lead printing FAIL. This way, users don't accidentally think that